+++ src/sagemaker/amazon/pca.py
@@ -64,6 +64,7 @@ class PCA(AmazonAlgorithmEstimatorBase):
         **kwargs
     ):
         """A Principal Components Analysis (PCA)
+
         :class:`~sagemaker.amazon.amazon_estimator.AmazonAlgorithmEstimatorBase`.
 
         This Estimator may be fit via calls to
@@ -132,8 +133,9 @@ class PCA(AmazonAlgorithmEstimatorBase):
         self.extra_components = extra_components
 
     def create_model(self, vpc_config_override=VPC_CONFIG_DEFAULT, **kwargs):
-        """Return a :class:`~sagemaker.amazon.pca.PCAModel` referencing the
-        latest s3 model data produced by this Estimator.
+        """Return a :class:`~sagemaker.amazon.pca.PCAModel`.
+
+        It references the latest s3 model data produced by this Estimator.
 
         Args:
             vpc_config_override (dict[str, list[str]]): Optional override for VpcConfig set on
@@ -207,6 +209,8 @@ class PCAPredictor(Predictor):
         deserializer=RecordDeserializer(),
     ):
         """
+        Initialization for PCAPredictor.
+
         Args:
             endpoint_name (str): Name of the Amazon SageMaker endpoint to which
                 requests are sent.
@@ -228,13 +232,16 @@ class PCAPredictor(Predictor):
 
 
 class PCAModel(Model):
-    """Reference PCA s3 model data. Calling
-    :meth:`~sagemaker.model.Model.deploy` creates an Endpoint and return a
+    """Reference PCA s3 model data.
+
+    Calling :meth:`~sagemaker.model.Model.deploy` creates an Endpoint and return a
     Predictor that transforms vectors to a lower-dimensional representation.
     """
 
     def __init__(self, model_data, role, sagemaker_session=None, **kwargs):
         """
+        Initialization for PCAModel.
+
         Args:
             model_data (str): The S3 location of a SageMaker model data
                 ``.tar.gz`` file.