+++ src/sagemaker/apiutils/_boto_functions.py
@@ -68,9 +68,7 @@ def from_boto(boto_dict, boto_name_to_member_name, member_name_to_type):
             api_type, is_collection = member_name_to_type[member_name]
             if is_collection:
                 if isinstance(boto_value, dict):
-                    member_value = {
-                        key: api_type.from_boto(value) for key, value in boto_value.items()
-                    }
+                    member_value = api_type.from_boto(boto_value)
                 else:
                     member_value = [api_type.from_boto(item) for item in boto_value]
             else: