+++ src/sagemaker/s3.py
@@ -23,8 +23,7 @@ logger = logging.getLogger("sagemaker")
 
 
 def parse_s3_url(url):
-    """Returns an (s3 bucket, key name/prefix) tuple from a url with an s3
-    scheme.
+    """Returns an (s3 bucket, key name/prefix) tuple from a url with an s3 scheme.
 
     Args:
         url (str):