+++ src/sagemaker/transformer.py
@@ -21,9 +21,7 @@ from sagemaker.utils import base_name_from_image, name_from_base
 
 
 class Transformer(object):
-    """A class for handling creating and interacting with Amazon SageMaker
-    transform jobs.
-    """
+    """A class for handling creating and interacting with Amazon SageMaker transform jobs."""
 
     def __init__(
         self,
@@ -303,8 +301,9 @@ class Transformer(object):
 
     @classmethod
     def _prepare_init_params_from_job_description(cls, job_details):
-        """Convert the transform job description to init params that can be
-        handled by the class constructor
+        """Convert the transform job description to init params.
+
+        It can be handled by the class constructor.
 
         Args:
             job_details (dict): the returned job details from a
@@ -349,20 +348,7 @@ class _TransformJob(_Job):
         experiment_config,
         model_client_config,
     ):
-        """
-        Args:
-            transformer:
-            data:
-            data_type:
-            content_type:
-            compression_type:
-            split_type:
-            input_filter:
-            output_filter:
-            join_source:
-            experiment_config:
-            model_client_config:
-        """
+        """Placeholder docstring"""
 
         transform_args = cls._get_transform_args(
             transformer,
@@ -396,20 +382,7 @@ class _TransformJob(_Job):
         experiment_config,
         model_client_config,
     ):
-        """
-        Args:
-            transformer:
-            data:
-            data_type:
-            content_type:
-            compression_type:
-            split_type:
-            input_filter:
-            output_filter:
-            join_source:
-            experiment_config:
-            model_client_config:
-        """
+        """Placeholder docstring"""
 
         config = _TransformJob._load_config(
             data, data_type, content_type, compression_type, split_type, transformer
@@ -448,15 +421,7 @@ class _TransformJob(_Job):
 
     @staticmethod
     def _load_config(data, data_type, content_type, compression_type, split_type, transformer):
-        """
-        Args:
-            data:
-            data_type:
-            content_type:
-            compression_type:
-            split_type:
-            transformer:
-        """
+        """Placeholder docstring"""
         input_config = _TransformJob._format_inputs_to_input_config(
             data, data_type, content_type, compression_type, split_type
         )
@@ -480,14 +445,7 @@ class _TransformJob(_Job):
 
     @staticmethod
     def _format_inputs_to_input_config(data, data_type, content_type, compression_type, split_type):
-        """
-        Args:
-            data:
-            data_type:
-            content_type:
-            compression_type:
-            split_type:
-        """
+        """Placeholder docstring"""
         config = {"DataSource": {"S3DataSource": {"S3DataType": data_type, "S3Uri": data}}}
 
         if content_type is not None:
@@ -503,13 +461,7 @@ class _TransformJob(_Job):
 
     @staticmethod
     def _prepare_output_config(s3_path, kms_key_id, assemble_with, accept):
-        """
-        Args:
-            s3_path:
-            kms_key_id:
-            assemble_with:
-            accept:
-        """
+        """Placeholder docstring"""
         config = super(_TransformJob, _TransformJob)._prepare_output_config(s3_path, kms_key_id)
 
         if assemble_with is not None:
@@ -522,12 +474,7 @@ class _TransformJob(_Job):
 
     @staticmethod
     def _prepare_resource_config(instance_count, instance_type, volume_kms_key):
-        """
-        Args:
-            instance_count:
-            instance_type:
-            volume_kms_key:
-        """
+        """Placeholder docstring"""
         config = {"InstanceCount": instance_count, "InstanceType": instance_type}
 
         if volume_kms_key is not None:
@@ -537,12 +484,7 @@ class _TransformJob(_Job):
 
     @staticmethod
     def _prepare_data_processing(input_filter, output_filter, join_source):
-        """
-        Args:
-            input_filter:
-            output_filter:
-            join_source:
-        """
+        """Placeholder docstring"""
         config = {}
 
         if input_filter is not None: